assignment/pyproject.toml
[project]
name = "exercise02"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv>=1.1.1",
]
[dependency-groups]
frontend = [
    "redis>=6.4.0",
    "streamlit>=1.50.0",
]
backend = [
    {include-group = "worker"},
    "fastapi>=0.118.0",
    "stability-sdk>=0.8.6",
    "uvicorn>=0.35.0",
]
worker = [
    "stability-sdk>=0.8.6",
    "redis>=6.4.0",
    "rq>=1.13.0",
    "requests>=2.31.0",
]
